do not send more than on worker to clear the heap
For ahead streams, in the end when the work queue is empty and only heap has
some remaining output to be transferred to the output queue, we should not send
more than worker to do it because anyway only one worker can work on the heap
at a time. Otherwise it leads to unnecessary workers not doing anything and
starving the single worker which is actually clearing the heap.