Add "no-charts" build option
In a build with the "dev" flag we build charting executable. However, "Charts"
package is a heavy dependency, takes lot of time to build and is not often up
to date with newer compilers making the cabal new-build fail. With this commit
we can use "no-charts" flag to disable charting to let the build proceed with
the "dev" flag.