Restrict maxThreads to maxBuffer
Currently maxThreads are independent of maxBuffer. If maxThreads are larger
than maxBuffer we may dispatch as many workers. Since we check the buffer limit
only after the worker yields a value, it may be too late and we may exceed the
buffer already. To make sure that we never exceed the buffer we restrict the
maxWorkers to maxBuffer value.

The invariant maxThreads <= maxBuffer holds.