Do not run each individual stream action with captured state
We already run the whole thread with captured monad state (in doFork). Running
with captured state in parallel stream is unnecessary. This leads to buildup of
memory on thread stack and memory keeps growing continuously. Also with the
removal of this call we are at 8 seconds vs 34 seconds in concatMapBy parallel
benchmark.