implement arraysOf in terms of chunksOf
* Reimplement groupsOf for better fusion
* Use A.writeNUnsafe for better fusion with groupsOf

Now we can use chunksOf/groupsOf to implement arraysOf with equivalent
performance. This commit hides the arraysOf from the Array module and export it
via Streamly.Internal because in one case (writeS) it still is faster. This can
be resolved once GHC issue is resolved. Even though writeS is not an officially
exported API, this case indicates that there may potentially be more such cases
which do not fuse well, so we are keeping this as a backup for such cases for
now.