improve perf of functor and applicative instances
Functor instances of various stream types are directly expressed in terms of
StreamD map impl.

Applicative instances now use concatMap instead of using "ap" from Monad.

Rewrite rules to specialize unfoldrM to WSerial and ZipSerial as well so
that it performs well for those. This was causing benchmarks for these
to be much slower beacause of unfoldrM used in the benchmarks being
slow.  For this, I had to move the zipWith* impls into the Zip module so
that we can import the ZipSerial type in Prelude.