Fix space leak in async combinator
We run the stream in a worker thread using captured monadic state from the
parent thread. The call to mrun caused memory to stack up becuase it is called
again and again and never returns. We need this call only once when a stream is
picked for execution from the work queue. We do not need to call mrun on every
yield. The call on yield is removed by this commit.