use -threaded, -N2, -M256M options for threaded benchmarks
Many issues are observed only with -threaded and -N2. We are also limiting
the heap memory to 256MB as all the benchmarks should ideally finish well below
that, there should be nothing that consumes a whole lot of memory. If we find
such genuine benchmarks we can make exceptions for those.