Remove the stream component from the Zipper
Earlier we had the whole input stream in the Zipper and when running
ParserK we could consume from the stream as we needed.  Now we drive
ParserK just like ParserD one element at a time in a loop. The stream
consumption now happens in this outer loop instead of inside parserK.

This makes the "parse" function in ParserK simpler and ParserD and
ParserK are more or less similarly driven.