Create a bench-runner wrapper that uses bench-config
Consolidate the specific customizations to bench-config.sh.

Integrate bench-exec-one.sh into the main bench.sh script. After the
gauge/tasty unification changes, we do not have a requirement for a
separate executable now.

Now we can push the generic part of the script to the benchmark/bench-report
package. And that package can be reused by streaming-benchmarks as well.
We can make that as a submodule.

Ultimately we should be converting these shell scripts to Haskell.