Add modular scanning functions for transformations (#1809)
Usable across folds and streams. Implemented common transformations
using scans. Performance of single operation benchmark seems to be
unaffected but in some cases combinations of multiple benchmarks or the
same operation multiple times is degraded.

The perf impact could perhaps be improved if we have a separate Scan
type where we do not need a pre-initialization of state.